refactor(plugin-assets): move code out of upload/list commands
